requirements.txt
python==3.10
pytorch==2.2.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
xgboost==3.0.0